+++ lib/src/runner/commands/add_repo_command.dart
@@ -5,7 +5,7 @@ import '../../commands/add_repo.dart';
 class AddRepoCommand extends Command<void> {
   AddRepoCommand() {
     argParser.addFlag(
-      'no_rest',
+      'no-rest',
       negatable: false,
       help: 'Skip generating the REST datasource and API paths.',
     );
@@ -18,7 +18,7 @@ class AddRepoCommand extends Command<void> {
   }
 
   @override
-  String get name => 'add_repo';
+  String get name => 'add-repo';
 
   @override
   String get description =>
@@ -26,16 +26,16 @@ class AddRepoCommand extends Command<void> {
 
   @override
   String get invocation =>
-      '${runner?.executableName} $name <feature> <repo_name> [--no_rest] [--add-sample]';
+      '${runner?.executableName} $name <feature> <repo_name> [--no-rest] [--add-sample]';
 
   @override
   String get usage =>
-      'Generate a repository interface (domain) and implementation (data).\n\nUsage: ${runner?.executableName} $name <feature> <repo_name> [--no_rest] [--add-sample]';
+      'Generate a repository interface (domain) and implementation (data).\n\nUsage: ${runner?.executableName} $name <feature> <repo_name> [--no-rest] [--add-sample]';
 
   @override
   void run() => addRepo(
     argResults!.rest,
-    noRest: argResults!['no_rest'] as bool,
+    noRest: argResults!['no-rest'] as bool,
     addSample: argResults!['add-sample'] as bool,
   );
 }